requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
filelock==3.13.1
fsspec==2024.6.1
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
packaging==25.0
pandas==2.3.1
pillow==11.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.6.1
setuptools==70.2.0
six==1.17.0
sympy==1.13.1
tokenizers==0.21.4
# --- PyTorch trio: wheels compatible with Python 3.11 & CUDA 12.1 ---
torch==2.5.1          # auto-selects +cu121 wheel on Colab
torchaudio==2.5.1
torchvision==0.20.1
# ---------------------------------------------------------------
tqdm==4.67.1
transformers==4.55.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
